move constants into subdirectory
