+++ polling_stations/settings/constants/tiles.py
